pyproject.toml
# SPDX-FileCopyrightText: 2025 2025 wahl.chat
#
# SPDX-License-Identifier: PolyForm-Noncommercial-1.0.0

[tool.poetry]
name = "wahl-chat-backend"
version = "1.0.0"
description = "The wahl.chat backend implements a political information chatbot that enables users to engage in a contemporary way with the positions of political parties and to receive answers to individual questions that can be substantiated with sources."
authors = ["wahl.chat <info@wahl.chat>"]
license = "PolyForm-Noncommercial-1.0.0"
readme = "README.md"
packages = [
    {include = "src"}
]
repository = "https://github.com/llm-psychology-group/wahl-chat-backend"

[tool.poetry.dependencies]
python = ">=3.11 <=3.12"
aiohttp = "~3.9.5"
aiohttp-cors = "~0.7.0"
aiohttp-pydantic = "~2.4.0"
firebase-admin = "~6.6.0"
ipykernel = "~6.29.5"
ipython = "~8.29.0"
jupyter-client = "~8.6.3"
jupyter-core = "~5.7.2"
langchain = "~0.3.7"
langchain-community = "~0.3.7"
langchain-core = "~0.3.19"
langchain-google-genai = "^2.0.8"
langchain-openai = "~0.2.9"
langchain-pinecone = "~0.2.0"
langchain-text-splitters = "~0.3.2"
openai = "~1.59.8"
pinecone-client = "~5.0.1"
pydantic = "~2.10.0"
pypdf = "~5.1.0"
pytest = "~8.3.3"
pytest-asyncio = "~0.24.0"
python-dotenv = "~1.0.1"
python-socketio = "~5.12.1"
pyyaml = "~6.0.2"
websocket-client = "1.8.0"
xxhash = "^3.5.0"

[tool.poetry.group.dev.dependencies]
mypy = "~1.12.1"
pre-commit = "~4.0.1"
ruff = "~0.8.6"

[tool.mypy]
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
